chore: CR improvements
