+++ .gitignore
@@ -2,4 +2,3 @@
 .env
 benches.json
 target
-node_modules
