+++ deno.jsonc
@@ -29,7 +29,6 @@
   },
   "tasks": {
     "run": "deno run -A",
-    "run:browser": "deno task run test_browser/gen_tests.ts && deno task run test_browser/test_browser.ts & node test_browser/puppeteer/puppeteer.js",
     "debug": "deno task run --inspect-brk",
     "udd": "deno task star && deno task run https://deno.land/x/udd@0.5.0/main.ts target/star.ts",
     "dnt": "deno task run _tasks/dnt.ts",
@@ -37,6 +36,8 @@
     "codegen": "deno task run cache.ts examples/mod.ts",
     "test": "deno task run test_util/ctx.ts -- deno test -A -L=info --ignore=target --parallel",
     "test:update": "deno task test -- -- --update",
+    "test:browser": "deno task run test_browser/gen_tests.ts && deno task run test_browser/server.ts & node test_browser/browser.js",
+    "test:browser:init": "deno cache test_browser/gen_tests.ts & deno cache test_browser/server.ts & npm install --prefix test_browser",
     "bench": "deno bench -A",
     "moderate": "deno task run https://deno.land/x/moderate@0.0.5/mod.ts && dprint fmt"
   }
