+++ test_browser/.gitignore
@@ -1,0 +1,3 @@
+node_modules
+package-lock.json
+# static -> uncomment when gen_tests.ts ready
