--- test_browser/expected_results.ts
@@ -1,8 +1,0 @@
-type IExpectedResults = Record<string, string>
-
-const derived =
-  "[1000,1001,2000,2002,2004,2006,2007,2008,2011,2012,2013,2019,2021,2026,2027,2030,2031,2032,2034,2035,2037,2039,2040,2043,2046,2048,2051,2052,2053,2055,2056,2058,2086,2090,2091,2092]"
-
-const derived2 = "{\"value\":\"Derived 2 test results\"}"
-
-export default { derived, derived2 } as IExpectedResults
