+++ test_browser/static/derived.html
@@ -3,15 +3,10 @@
   import * as C from "./capi.js"
 
   ;(async function browserTest() {
-    const root = C.entryRead(C.polkadot)("Paras", "Parachains", [])
-    .access("value")
+    const root = C.entryRead(C.polkadot)("Paras", "Parachains", []).access("value")
 
-    const result = await root.run()
+    console.log(C.throwIfError(await root.run()))
 
-    await fetch('http://localhost:8000/derived', {
-      method: "POST",
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify(result)
-    })
+    console.info('derived')
   })()
 </script>
